docs/requirements.txt
Faker
python-docx
XlsxWriter
python-pptx
Pillow
google-genai